+++ src/app/actions/index.ts
@@ -1,4 +1,5 @@
 import { Action as ReduxAction } from 'redux';
+import uuid from 'uuid/v4';
 
 import {
   ADD_TODO,
@@ -9,6 +10,8 @@ import {
   FETCH_TODOS_REJECTED,
 } from '../constants';
 
+import { NormalizedTodoList } from '../models';
+
 export interface Action<P = undefined, M = undefined> extends ReduxAction {
   readonly type: string;
   readonly payload?: P;
@@ -17,11 +20,13 @@ export interface Action<P = undefined, M = undefined> extends ReduxAction {
 }
 
 export interface AddTodoPayload {
+  readonly id: string;
   readonly text: string;
+  readonly completed: false;
 }
 
 export interface ToggleTodoPayload {
-  readonly id: number;
+  readonly id: string;
 }
 
 export interface AddTodoAction extends Action<AddTodoPayload>  {
@@ -40,7 +45,7 @@ export interface FetchTodosPendingAction extends Action {
   type: FETCH_TODOS_PENDING;
 }
 
-export interface FetchTodosFulfilledAction extends Action<string[]> {
+export interface FetchTodosFulfilledAction extends Action<NormalizedTodoList> {
   type: FETCH_TODOS_FULFILLED;
 }
 
@@ -51,11 +56,13 @@ export interface FetchTodosRejectedAction extends Action<Error> {
 export const addTodo: (text: string) => AddTodoAction = text => ({
   type: ADD_TODO,
   payload: {
+    id: uuid(),
     text,
+    completed: false,
   },
 });
 
-export const toggleTodo: (id: number) => ToggleTodoAction = id => ({
+export const toggleTodo: (id: string) => ToggleTodoAction = id => ({
   type: TOGGLE_TODO,
   payload: {
     id,
@@ -70,10 +77,11 @@ export const fetchTodosPending: () => FetchTodosPendingAction = () => ({
   type: FETCH_TODOS_PENDING,
 });
 
-export const fetchTodosFulfilled: (todos: string[]) => FetchTodosFulfilledAction = todos => ({
-  type: FETCH_TODOS_FULFILLED,
-  payload: todos,
-});
+export const fetchTodosFulfilled: (todos: NormalizedTodoList) => FetchTodosFulfilledAction =
+  todos => ({
+    type: FETCH_TODOS_FULFILLED,
+    payload: todos,
+  });
 
 export const fetchTodosRejected: (error: Error) => FetchTodosRejectedAction = error => ({
   type: FETCH_TODOS_REJECTED,