+++ src/app/components/TodoList/TodoList.tsx
@@ -8,7 +8,7 @@ const styles = require('./TodoList.scss');
 
 export interface TodoListProps {
   todos: Todo[];
-  onTodoClick: (id: number) => void;
+  onTodoClick: (id: string) => void;
 }
 
 const TodoList: React.SFC<TodoListProps> = ({ todos, onTodoClick }) => {