+++ src/app/reducers/index.ts
@@ -1,14 +1,13 @@
 import { combineReducers } from 'redux';
 
-import { todosReducer as todos } from './todos';
-import { Todo } from '../models';
+import { todosReducer as todos, TodosState } from './todos';
 
 const rootReducer = combineReducers<State>({
   todos,
 });
 
 export interface State {
-  todos?: Todo[];
+  todos?: TodosState;
 }
 
 export default rootReducer;