+++ src/app/reducers/todos.ts
@@ -4,41 +4,60 @@ import { AddTodoAction, ToggleTodoAction, FetchTodosFulfilledAction } from '../a
 
 type TodosAction = AddTodoAction | ToggleTodoAction | FetchTodosFulfilledAction;
 
-export const INITIAL_STATE: Todo[] = [];
+export interface TodosState {
+  byId: {
+    [id: string]: Todo;
+  };
+  allIds: string[];
+}
 
-let nextTodoId = 0;
+export const INITIAL_STATE: TodosState = {
+  byId: {},
+  allIds: [],
+};
 
-const todoFactory: (text: string) => Todo = text => ({
-  // tslint:disable-next-line:no-increment-decrement
-  id: nextTodoId++,
-  text,
-  completed: false,
-});
-
-export const todosReducer = (todos: Todo[] = INITIAL_STATE, action: TodosAction) => {
+export const todosReducer = (prevState: TodosState = INITIAL_STATE, action: TodosAction) => {
   switch (action.type) {
     case ADD_TODO:
-      return [
-        ...todos,
-        todoFactory(action.payload.text),
-      ];
+      return {
+        ...prevState,
+        byId: {
+          ...prevState.byId,
+          [action.payload.id]: action.payload,
+        },
+        allIds: [
+          ...prevState.allIds,
+          action.payload.id,
+        ],
+      };
     case TOGGLE_TODO:
-      return todos.map((todo) => {
-        if (todo.id !== action.payload.id) {
-          return todo;
-        }
+      if (!prevState.byId[action.payload.id]) {
+        return prevState;
+      }
 
-        return {
-          ...todo,
-          completed: !todo.completed,
-        };
-      });
+      return {
+        ...prevState,
+        byId: {
+          ...prevState.byId,
+          [action.payload.id]: {
+            ...prevState.byId[action.payload.id],
+            completed: !prevState.byId[action.payload.id].completed,
+          },
+        },
+      };
     case FETCH_TODOS_FULFILLED:
-      return [
-        ...todos,
-        ...action.payload.map(todo => todoFactory(todo)),
-      ];
+      return {
+        ...prevState,
+        byId: {
+          ...prevState.byId,
+          ...action.payload.entities.todos,
+        },
+        allIds: [
+          ...prevState.allIds,
+          ...action.payload.result,
+        ],
+      };
     default:
-      return todos;
+      return prevState;
   }
 };