+++ src/app/selectors/index.ts
@@ -1,3 +1,3 @@
 import { State } from '../reducers';
 
-export const getTodos = (state: State) => state.todos;
+export const getTodos = (state: State) => state.todos.allIds.map(id => state.todos.byId[id]);